Predict hotkey and slot button presses
